+++ js/CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 0.2.2
   * Improved the look and working of the Load spreadsheet button. Load Default now loads
   the default data file.
+  * Added a clear button for the input text area.
 
 ## Version 0.2.1
   * The web page Default Output languages list is now auto-populated based on the actual